feature: Add TutorLessonResponse screen with needed components and updates.
Add a component DisplayLessonDetails that can be reused when a tutor need to accept a lesson after a match has been found.
Update the LessonStatus enum to separate pending lesson for tutor and for student.
Update the view models with update function of the lesson and get a profile by uid.