fix: Fix goBack action in TutorMatchingScreen
When we goBack in navigate to the last screen so the screen that add a lesson.
This commit fix the bug that when we goBack we need to create a new Lesson and we loose the lesson we had before 8even if stored in Firebase.
AddLesson screen now check if a lesson is selected and display this lesson if any.